Document-Eligibility-Agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-eligibility-agent"
version = "0.1.0"
description = "AI-powered document processing agent for NY State eligibility verification"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "NY State AI Hackathon Team"}
]
keywords = ["ai", "document-processing", "eligibility", "azure", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Government",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "flask>=3.1.0",
    "flask-cors>=4.0.0",
    "azure-ai-documentintelligence>=1.0.0",
    "azure-ai-projects>=1.0.0",
    "azure-storage-blob>=12.0.0",
    "azure-cosmos>=4.5.0",
    "msgraph-sdk>=1.0.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

[project.scripts]
doc-eligibility = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
